+++ rubato/struct/gameobject/sprites/__init__.py
@@ -1,5 +1,4 @@
 """Holds all image-related components."""
-from .raster import Raster
-from .image import Image
+from .raster import Raster, Image
 from .animation import Animation
 from .spritesheet import Spritesheet